requirements.txt
pandas
plotly-express
streamlit
matplotlib
nbformat
jupyter 
ipykernel